+++ VC_firmware/VC_firmware/src/interface.h
@@ -6,8 +6,8 @@
 #include "config.h"
 
 void send_message(String channel, String mssg, float data);
-int validator(float &user_height, float &user_period, String &input);
-int check_height(float &num);
-int check_period(float &num);
+int validator(uint16_t &user_height, uint16_t &user_period, String &input);
+int check_height(uint16_t &num);
+int check_period(uint16_t &num);
 String get_input();
-int listener(float &user_height, float &user_period);
+int listener(uint16_t &user_height, uint16_t &user_period);
